requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-log-request-id==1.6.0
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
psycopg2==2.7.7
psycopg2-binary==2.8.5
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
waitress==1.4.4
whitenoise==5.2.0
